quick commit

+++ lgal_convert/gen_lgal_trees.c
@@ -43,7 +43,7 @@ int main(int argc, char **argv) {
     for(j=0;j<BOX_DIVISIONS;j++)
       for(k=0;k<BOX_DIVISIONS;k++) {
 	findex = i*BOX_DIVISIONS*BOX_DIVISIONS+j*BOX_DIVISIONS+k;
-	sprintf(buffer,"tree_%d.dat",i);
+	sprintf(buffer,"tree_%d.dat",(int)i);
 	printf("reading %s\n",buffer);
 	/* read_tree(treefile); */
 	/* build_lgal_tree(); */
